Updated Patient.java
Decided an ID would be a better way to pass info around rather than the name. Adjustments in doctor will be made to accommodate. Also removed the symptom dependency. Now they will be strings that can be used to retrieve data from other classes instead of storing it directly in the patient object.