Reworked the data storage and retrieval system
Created a file parser since it was just repeated code. Simplifies the specialization and symptoms classes. There is currently an issue with the retrieval. I am confident it is a parsing issue, but I can go through the file parser tomorrow or Monday. The issue is that it seems to be saving null values (probably the breaklines).