Massive overhaul of data management
Removed Department because it was unimportant. I have bridged the Symptom/Specialization and FileParser classes with a Map class for them both. It is a singleton so the first time it is called it will parse all the information and be available to any other Symptom or Specialization without having to go through the whole file again. I will need to add some conditional logic in case someone tries to add a symptom or specialization that does not exist, but for now this should run.